@@ -1,10 +1,10 @@
 name: flutter_screen_orientation
 description: 实时监测Android/IOS屏幕方向的插件
-version: 0.0.1+1
+version: 1.0.1+1
 homepage: https://github.com/515445681/flutter_screen_orientation
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
